+++ lib/ui/wellness/rings/WellnessRingSelectPredefinedPanel.dart
@@ -166,7 +166,7 @@ class _WellnessRingButton extends StatefulWidget{
   final void Function(BuildContext context) onTapWidget;
   final void Function(BuildContext context)? onTapRightWidget;
 
-  // ignore: unused_element
+  // ignore: unused_element_parameter
   const _WellnessRingButton({Key? key, required this.label, this.description, required this.onTapWidget, this.onTapRightWidget, required this.toggled}) : super(key: key);
 
   @override