+++ lib/ui/widgets/QrCodePanel.dart
@@ -29,7 +29,7 @@ import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
 import 'package:rokwire_plugin/utils/image_utils.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
-import 'package:share/share.dart';
+import 'package:share_plus/share_plus.dart';
 
 class QrCodePanel extends StatefulWidget with AnalyticsInfo { //TBD localize
   //final Event2? event;
@@ -356,8 +356,7 @@ class _QrCodePanelState extends State<QrCodePanel> {
     File capturedFile = File(fullPath);
     await capturedFile.writeAsString(widget.digitalCardShare ?? '');
     if (mounted) {
-      Share.shareFiles([fullPath],
-        mimeTypes: ['text/vcard'],
+      Share.shareXFiles([XFile(fullPath, mimeType: 'text/vcard',)],
         text: widget.saveWatermarkText,
       );
     }