+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Use the common method for launching urls [#4898](https://github.com/rokwire/illinois-app/issues/4898).
+### Changed
+- Upgrade to Flutter 3.29.2, upgraded plugins [#4899](https://github.com/rokwire/illinois-app/issues/4899).
 
 ## [6.2.2] - 2025-03-14
 ### Changed